requirements.txt
Pillow==10.1.0
numpy==1.26.2
